+++ src/sqlfmt/token.py
@@ -6,6 +6,7 @@ class TokenType(Enum):
     FMT_OFF = auto()
     FMT_ON = auto()
     JINJA = auto()
+    JINJA_COMMENT = auto()
     JINJA_START = auto()
     JINJA_END = auto()
     QUOTED_NAME = auto()