+++ tests/data/unformatted/101_multiline.sql
@@ -54,16 +54,12 @@ select * from renamed /* what!?! */ where true
  * And even /* some {% special characters %}
  * but we're not going to parse those
 */
-
 with
     source as (select * from {{ ref('my_model') }}),
-    renamed as (
-        /* This is a multiline comment in very bad style,
+    /* This is a multiline comment in very bad style,
     * which starts and ends on lines with other tokens.
     */
-        select id, another_field, and_another, and_still_another
-        from source
-    ),
+    renamed as (select id, another_field, and_another, and_still_another from source),
     {% set my_variable_in_bad_style = [
         "a",
         "short",
@@ -72,12 +68,11 @@ with
         "strings"
     ] %}
     
-    {#
+{#
  # And this is a nice multiline jinja comment
  # that we will also handle.
 #}
-    
+/* what!?! */
 select *
-from renamed /* what!?! */
+from renamed
 where true
-