+++ tests/data/unit_tests/test_splitter/test_simple_comment_split.sql
@@ -7,16 +7,11 @@ where -- this should stay
     true
 -- one last comment
 )))))__SQLFMT_OUTPUT__(((((
--- not distinct, just an ordinary select here, no big deal at all, it's okay really
 select
-    -- here is a long comment to be wrapped above this line
     my_first_field + an_operation,
-    my_second_field, -- a short comment
-    -- here is another long comment to be wrapped but not indented
+    my_second_field,
     my_third_field + another_long_operation
--- another comment that is a little bit too long to stay here
 from
     my_really_long_data_source
-where -- this should stay
+where
     true
-    -- one last comment